ml-services/requirements.txt
# Core Flask dependencies
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Limiter==2.7.0

# Machine Learning and Data Processing
scikit-learn==1.3.2
numpy==1.26.2
joblib==1.3.2

# Text Processing and NLP
nltk==3.8.1

# System Monitoring and Utilities (optional for production)
psutil==5.9.6
colorlog==6.8.0
colorama==0.4.6

# HTTP and Request Handling
requests==2.31.0
urllib3==2.1.0

# Environment and Configuration
python-dotenv==1.0.0

# Security and Validation
werkzeug==3.0.1

# Data Generation (for training data)
Faker==20.1.0

# Production server (alternative to Flask dev server)
gunicorn==21.2.0

# Memory optimization
memory-profiler==0.61.0